Slight refactoring of load_module_from_path() function to improve readability plus one bugfix
Mosly disambiguated 1- and 2-letter variable names into longer "talking" names,
swapped long concatenations when constructing error responses for a % substitution operator,
converted a couple of if-return if-return into elifs,
and converted a nested "if-if" into "if-and".

The only real bug discovered was that  "getattr(c, 'init')"  which was supposed to check the existence of init() method
actually fails to do so (this was masked by checking the other condition first).
So I swapped it for a lookup in dir(module_code) list.